[2021] Standardize on int64_t

+++ 2021/day01/lib.cc
@@ -4,16 +4,16 @@
 
 #include <limits>
 
-uint32_t day01::count_increases(const std::vector<std::string> &input, size_t window_size)
+int64_t day01::count_increases(const std::vector<std::string> &input, size_t window_size)
 {
-    uint32_t increases = 0;
+    int64_t increases = 0;
 
     auto values = common::vector_parse_int(input);
-    int32_t previous = std::numeric_limits<int32_t>::max();
+    int64_t previous = std::numeric_limits<int64_t>::max();
 
     for (size_t i = window_size - 1; i < input.size(); ++i)
     {
-        int sum = 0;
+        int64_t sum = 0;
         for (size_t j = 0; j < window_size; ++j)
         {
             sum += values[i - j];
