+++ 2021/day01/lib.h
@@ -6,7 +6,7 @@
 
 namespace day01
 {
-    uint32_t count_increases(const std::vector<std::string> &input, size_t window_size);
+    int64_t count_increases(const std::vector<std::string> &input, size_t window_size);
 }
 
 #endif
