+++ 2021/libcommon/include/common/vector_convert.h
@@ -6,7 +6,7 @@
 
 namespace common
 {
-    std::vector<int32_t> vector_parse_int(const std::vector<std::string> &input);
+    std::vector<int64_t> vector_parse_int(const std::vector<std::string> &input);
 }
 
 #endif
